+++ tencentcloud/resource_tc_cam_role_test.go
@@ -53,7 +53,7 @@ func testAccCheckCamRoleDestroy(s *terraform.State) error {
 
 		instance, err := camService.DescribeRoleById(ctx, rs.Primary.ID)
 		if err == nil && instance != nil {
-			return fmt.Errorf("[TECENT_TERRAFORM_CHECK][CAM role][Destroy] check: CAM role still exists: %s", rs.Primary.ID)
+			return fmt.Errorf("[CHECK][CAM role][Destroy] check: CAM role still exists: %s", rs.Primary.ID)
 		}
 	}
 	return nil
@@ -66,10 +66,10 @@ func testAccCheckCamRoleExists(n string) resource.TestCheckFunc {
 
 		rs, ok := s.RootModule().Resources[n]
 		if !ok {
-			return fmt.Errorf("[TECENT_TERRAFORM_CHECK][CAM role][Exists] check: CAM role %s is not found", n)
+			return fmt.Errorf("[CHECK][CAM role][Exists] check: CAM role %s is not found", n)
 		}
 		if rs.Primary.ID == "" {
-			return fmt.Errorf("[TECENT_TERRAFORM_CHECK][CAM role][Exists] check: CAM role id is not set")
+			return fmt.Errorf("[CHECK][CAM role][Exists] check: CAM role id is not set")
 		}
 		camService := CamService{
 			client: testAccProvider.Meta().(*TencentCloudClient).apiV3Conn,
@@ -79,7 +79,7 @@ func testAccCheckCamRoleExists(n string) resource.TestCheckFunc {
 			return err
 		}
 		if instance == nil {
-			return fmt.Errorf("[TECENT_TERRAFORM_CHECK][CAM role][Exists] check: CAM role %s is not exist", rs.Primary.ID)
+			return fmt.Errorf("[CHECK][CAM role][Exists] check: CAM role %s is not exist", rs.Primary.ID)
 		}
 		return nil
 	}