+++ tencentcloud/resource_tc_cam_user_test.go
@@ -124,7 +124,7 @@ func testAccCheckCamUserDestroy(s *terraform.State) error {
 
 		instance, err := camService.DescribeUserById(ctx, rs.Primary.ID)
 		if err == nil && (instance != nil && instance.Response != nil && instance.Response.Uid != nil) {
-			return fmt.Errorf("[TECENT_TERRAFORM_CHECK][CAM user][Destroy] check: CAM user still exists: %s", rs.Primary.ID)
+			return fmt.Errorf("[CHECK][CAM user][Destroy] check: CAM user still exists: %s", rs.Primary.ID)
 		}
 	}
 	return nil
@@ -137,10 +137,10 @@ func testAccCheckCamUserExists(n string) resource.TestCheckFunc {
 
 		rs, ok := s.RootModule().Resources[n]
 		if !ok {
-			return fmt.Errorf("[TECENT_TERRAFORM_CHECK][CAM user][Exists] check: CAM user %s is not found", n)
+			return fmt.Errorf("[CHECK][CAM user][Exists] check: CAM user %s is not found", n)
 		}
 		if rs.Primary.ID == "" {
-			return fmt.Errorf("[TECENT_TERRAFORM_CHECK][CAM user][Exists] check: CAM user id is not set")
+			return fmt.Errorf("[CHECK][CAM user][Exists] check: CAM user id is not set")
 		}
 		camService := CamService{
 			client: testAccProvider.Meta().(*TencentCloudClient).apiV3Conn,
@@ -150,7 +150,7 @@ func testAccCheckCamUserExists(n string) resource.TestCheckFunc {
 			return err
 		}
 		if instance == nil || instance.Response == nil || instance.Response.Uid == nil {
-			return fmt.Errorf("[TECENT_TERRAFORM_CHECK][CAM user][Exists] check: CAM user %s is not exist", rs.Primary.ID)
+			return fmt.Errorf("[CHECK][CAM user][Exists] check: CAM user %s is not exist", rs.Primary.ID)
 		}
 		return nil
 	}